requirements.txt
mariadb==1.1.7
schedule==1.2.0
python-telegram-bot
python-dotenv==1.0.0
Babel==2.12.1
#Telethon==1.29.2